routing fix + re-map URLs 06
